Refactored ASE code into own class
