+++ src/main/java/legend/game/submap/SMap.java
@@ -317,17 +317,8 @@ public class SMap extends EngineState {
 
   private final Model124 submapModel_800d4bf8 = new Model124("Submap");
 
-  private final Model124 tmdDustModel_800d4d40 = new Model124("Dust");
-
-  private final OrthoTrailParticle54 orthoQuadTrail = new OrthoTrailParticle54();
-
-  private final TmdTrailParticle20 tmdTrail_800d4ec0 = new TmdTrailParticle20();
 
   private final SmokeParticleEffect smokeCloudEffect_800d4f50 = new SmokeParticleEffect();
-
-  private final List<LawPodTrailSegment34> lawPodTrail_800d4f90 = new ArrayList<>();
-
-  private final List<TrailSegmentVertices14> lawPodTrailVerts_800d4fd0 = new ArrayList<>();
 
   private final TriangleIndicator44[] _800d4ff0 = new TriangleIndicator44[21];
   {
@@ -410,26 +401,6 @@ public class SMap extends EngineState {
 
   private final Rect4i _800d6b48 = new Rect4i(576, 368, 16, 1);
 
-  private final Vector3f[] footprintQuadVertices_800d6b7c = {
-    new Vector3f(-10.0f, 0.0f, -22.0f),
-    new Vector3f( 10.0f, 0.0f, -22.0f),
-    new Vector3f(-10.0f, 0.0f,  22.0f),
-    new Vector3f( 10.0f, 0.0f,  22.0f),
-    new Vector3f(-12.0f, 0.0f, - 8.0f),
-    new Vector3f(- 2.0f, 0.0f, - 8.0f),
-    new Vector3f(-12.0f, 0.0f,   8.0f),
-    new Vector3f(- 2.0f, 0.0f,   8.0f),
-    new Vector3f(  2.0f, 0.0f, - 8.0f),
-    new Vector3f( 12.0f, 0.0f, - 8.0f),
-    new Vector3f(  2.0f, 0.0f,   8.0f),
-    new Vector3f( 12.0f, 0.0f,   8.0f),
-  };
-  private final int[] orthoDustUs_800d6bdc = {96, 112, 64, 0};
-  private final int[] dustTextureWidths_800d6bec = {15, 15, 31, 23};
-  private final int[] dustTextureHeights_800d6bfc = {31, 31, 31, 23};
-  private final int[] particleFadeDelay_800d6c0c = {120, 0, 120};
-
-
   private final Vector3f savePointV0_800d6c28 = new Vector3f(-24.0f, -32.0f,  24.0f);
   private final Vector3f savePointV1_800d6c30 = new Vector3f( 24.0f, -32.0f,  24.0f);
   private final Vector3f savePointV2_800d6c38 = new Vector3f(-24.0f, -32.0f, -24.0f);
@@ -454,8 +425,6 @@ public class SMap extends EngineState {
   private AnmFile savepointAnm2;
   private CContainer savepointTmd;
   private TmdAnimationFile savepointAnimation;
-  private CContainer dustTmd;
-  private TmdAnimationFile dustAnimation;
 
   private final String bigArrow_800d7620 = "big_arrow.tim";
   private final String smallArrow_800d7c60 = "small_arrow.tim";
@@ -751,9 +720,6 @@ public class SMap extends EngineState {
   private final Vector2i clut_800f9e5e = new Vector2i();
   private int snowState_800f9e60;
 
-  private int lawPodTrailSegmentCount_800f9e78;
-
-  private final LawPodTrailData18[] lawPodTrailSegments_800f9e7c = new LawPodTrailData18[8];
   private int momentaryIndicatorTicks_800f9e9c;
 
   private int _800f9ea0;
@@ -783,6 +749,8 @@ public class SMap extends EngineState {
   };
 
   private final MapIndicator mapIndicator = new MapIndicator();
+
+  private final AttachedSobjEffect attachedSobjEffect = new AttachedSobjEffect();
 
   @Override
   public Function<RunningScript, FlowControl>[] getScriptFunctions() {
@@ -1367,6 +1335,7 @@ public class SMap extends EngineState {
     partCoord.flg--;
   }
 
+  // TODO Clean this up
   @Method(0x800daa3cL)
   private void renderSmapModel(final Model124 model) {
     zOffset_1f8003e8 = model.zOffset_a0;
@@ -3091,7 +3060,7 @@ public class SMap extends EngineState {
       }
 
       //LAB_800e1390
-      this.tickAttachedSobjEffects(sobj.model_00, sobj.attachedEffectData_1d0);
+      this.attachedSobjEffect.tickAttachedSobjEffects(sobj.model_00, sobj.attachedEffectData_1d0, this.unloadSubmapParticles_800c6870, this.screenOffsetX_800cb568, this.screenOffsetY_800cb56c);
     }
 
     //LAB_800e139c
@@ -4080,7 +4049,7 @@ public class SMap extends EngineState {
 
   @Method(0x800e4708L)
   private void renderSubmap() {
-    this.renderAttachedSobjEffects();
+    this.attachedSobjEffect.renderAttachedSobjEffects(this.screenOffsetX_800cb568, this.screenOffsetY_800cb56c);
     this.renderSubmapOverlays();
     this.handleAndRenderSubmapEffects();
     applyModelRotationAndScale(this.playerModel_800c6748);
@@ -4581,8 +4550,8 @@ public class SMap extends EngineState {
           this.savepointAnm2 = new AnmFile(files.get(1));
           this.savepointTmd = new CContainer("Savepoint", files.get(2));
           this.savepointAnimation = new TmdAnimationFile(files.get(3));
-          this.dustTmd = new CContainer("Dust", files.get(4));
-          this.dustAnimation = new TmdAnimationFile(files.get(5));
+          this.attachedSobjEffect.dustTmd = new CContainer("Dust", files.get(4));
+          this.attachedSobjEffect.dustAnimation = new TmdAnimationFile(files.get(5));
         });
         this.smapLoadingStage_800cb430 = SubmapState.WAIT_FOR_NEWROOT_2;
       }
@@ -6182,322 +6151,6 @@ public class SMap extends EngineState {
     GPU.uploadData15(imageRect, imageAddress);
   }
 
-  @Method(0x800ef0f8L)
-  private void tickAttachedSobjEffects(final Model124 model, final AttachedSobjEffectData40 data) {
-    if(!flEq(data.transfer_1e.x, model.coord2_14.coord.transfer.x) || !flEq(data.transfer_1e.y, model.coord2_14.coord.transfer.y) || !flEq(data.transfer_1e.z, model.coord2_14.coord.transfer.z)) {
-      //LAB_800ef154
-      if(data.shouldRenderTmdDust_04) {
-        if(data.tick_00 % (data.instantiationIntervalDust30 * (3 - vsyncMode_8007a3b8)) == 0) {
-          final TmdTrailParticle20 inst = this.addTmdDustParticle(this.tmdTrail_800d4ec0);
-          inst.tick_00 = 0;
-          inst.maxTicks_18 = data.maxTicks_38;
-
-          final int size = data.size_28;
-          if(size < 0) {
-            inst.size_08 = -size;
-            inst.stepSize_04 = -inst.stepSize_04 / 20.0f;
-          } else if(size > 0) {
-            //LAB_800ef1e0
-            inst.size_08 = 0.0f;
-            inst.stepSize_04 = size / 20.0f;
-          } else {
-            //LAB_800ef214
-            inst.size_08 = 0.0f;
-            inst.stepSize_04 = 0.0f;
-          }
-
-          //LAB_800ef21c
-          inst.transfer.set(model.coord2_14.coord.transfer);
-          data.transfer_1e.set(model.coord2_14.coord.transfer);
-        }
-      }
-
-      //LAB_800ef240
-      if(data.shouldRenderFootprints_08) {
-        if(data.tick_00 % (data.instantiationIntervalFootprints_34 * (3 - vsyncMode_8007a3b8)) == 0) {
-          //LAB_800ef394
-          final OrthoTrailParticle54 inst = this.addOrthoQuadTrailParticle(this.orthoQuadTrail);
-
-          if(data.footprintMode_10 != 0) {
-            //LAB_800ef3e8
-            inst.renderMode_00 = 2;
-            inst.textureIndex_02 = 3;
-
-            //LAB_800ef3f8
-          } else {
-            inst.renderMode_00 = 0;
-            inst.textureIndex_02 = data.textureIndexType1_1c;
-
-            data.textureIndexType1_1c ^= 1;
-          }
-
-          //LAB_800ef3fc
-          inst.x_18 = this.screenOffsetX_800cb568;
-          inst.y_1c = this.screenOffsetY_800cb56c;
-          inst.tick_04 = 0;
-          inst.maxTicks_06 = 150;
-
-          final MV ls = new MV();
-          GsGetLs(model.coord2_14, ls);
-          GTE.setTransforms(ls);
-
-          final int textureIndex = inst.textureIndex_02;
-          if(textureIndex == 0) {
-            //LAB_800ef4b4
-            inst.z_4c = RotTransPers4(this.footprintQuadVertices_800d6b7c[4], this.footprintQuadVertices_800d6b7c[5], this.footprintQuadVertices_800d6b7c[6], this.footprintQuadVertices_800d6b7c[7], inst.sxy0_20, inst.sxy1_28, inst.sxy2_30, inst.sxy3_38);
-          } else if(textureIndex == 1) {
-            //LAB_800ef484
-            //LAB_800ef4b4
-            inst.z_4c = RotTransPers4(this.footprintQuadVertices_800d6b7c[8], this.footprintQuadVertices_800d6b7c[9], this.footprintQuadVertices_800d6b7c[10], this.footprintQuadVertices_800d6b7c[11], inst.sxy0_20, inst.sxy1_28, inst.sxy2_30, inst.sxy3_38);
-          } else if(textureIndex == 3) {
-            //LAB_800ef4a0
-            //LAB_800ef4b4
-            inst.z_4c = RotTransPers4(this.footprintQuadVertices_800d6b7c[0], this.footprintQuadVertices_800d6b7c[1], this.footprintQuadVertices_800d6b7c[2], this.footprintQuadVertices_800d6b7c[3], inst.sxy0_20, inst.sxy1_28, inst.sxy2_30, inst.sxy3_38);
-          }
-
-          //LAB_800ef4ec
-          if(inst.z_4c < 41) {
-            inst.z_4c = 41;
-          }
-
-          //LAB_800ef504
-          inst.stepBrightness_40 = 0.5f / 30;
-          inst.brightness_48 = 0.5f;
-          data.transfer_1e.set(model.coord2_14.coord.transfer);
-        }
-      }
-
-      //LAB_800ef520
-      if(data.shouldRenderOrthoDust_0c) {
-        if(data.tick_00 % (data.instantiationIntervalDust30 * (3 - vsyncMode_8007a3b8)) == 0) {
-          final OrthoTrailParticle54 inst = this.addOrthoQuadTrailParticle(this.orthoQuadTrail);
-          inst.renderMode_00 = 1;
-          inst.textureIndex_02 = 2;
-          inst.x_18 = this.screenOffsetX_800cb568;
-          inst.y_1c = this.screenOffsetY_800cb56c;
-
-          final Vector3f vert0 = new Vector3f(-data.size_28, 0.0f, -data.size_28);
-          final Vector3f vert1 = new Vector3f( data.size_28, 0.0f, -data.size_28);
-          final Vector3f vert2 = new Vector3f(-data.size_28, 0.0f,  data.size_28);
-          final Vector3f vert3 = new Vector3f( data.size_28, 0.0f,  data.size_28);
-
-          inst.tick_04 = 0;
-          inst.maxTicks_06 = (short)data.maxTicks_38;
-
-          final MV ls = new MV();
-          GsGetLs(model.coord2_14, ls);
-          GTE.setTransforms(ls);
-
-          //TODO The real code actually passes the same reference for sxyz 1 and 2, is that a bug?
-          inst.z_4c = RotTransPers4(vert0, vert1, vert2, vert3, inst.sxy0_20, inst.sxy1_28, inst.sxy2_30, inst.sxy3_38);
-
-          if(inst.z_4c < 41) {
-            inst.z_4c = 41;
-          }
-
-          //LAB_800ef6a0
-          final float halfSize = (inst.sxy3_38.x - inst.sxy0_20.x) / 2.0f;
-          inst.size_08 = halfSize;
-          inst.sizeStep_0c = halfSize / data.maxTicks_38;
-
-          inst.z0_26 = (inst.sxy3_38.x + inst.sxy0_20.x) / 2.0f;
-          inst.z1_2e = (inst.sxy3_38.y + inst.sxy0_20.y) / 2.0f;
-
-          inst.stepBrightness_40 = 0.5f / data.maxTicks_38;
-          inst.brightness_48 = 0.5f;
-          data.transfer_1e.set(model.coord2_14.coord.transfer);
-        }
-      }
-
-      //LAB_800ef728
-      if(data.shouldRenderLawPodTrail_18) {
-        if(!this.unloadSubmapParticles_800c6870) {
-          this.tickLawPodTrail(model, data);
-        }
-      }
-    }
-
-    //LAB_800ef750
-    data.tick_00++;
-  }
-
-  @Method(0x800ef798L)
-  private void renderTmdTrail() {
-    TmdTrailParticle20 prev = this.tmdTrail_800d4ec0;
-    TmdTrailParticle20 inst = prev.next_1c;
-
-    //LAB_800ef7c8
-    while(inst != null) {
-      if(inst.tick_00 >= inst.maxTicks_18 * (3 - vsyncMode_8007a3b8)) {
-        prev.next_1c = inst.next_1c;
-        inst = prev.next_1c;
-      } else {
-        //LAB_800ef804
-        inst.transfer.y -= 1.0f / (3 - vsyncMode_8007a3b8);
-
-        this.tmdDustModel_800d4d40.coord2_14.coord.transfer.set(inst.transfer);
-
-        // In retail, the shrinking tail of the snow slide snow cloud animation occurs because size_08 is in .12
-        // and RotMatrixXyz writes everything as shorts, which turns scale negative above 0x7fff (>=8.0f).
-        // To implement this more smoothly, stepSize_04 is switched to negative once size_08 reaches 8.0f.
-        if(inst.size_08 >= 8.0f) {
-          inst.stepSize_04 = -inst.stepSize_04;
-        }
-        inst.size_08 += inst.stepSize_04 / (3 - vsyncMode_8007a3b8);
-
-        this.tmdDustModel_800d4d40.coord2_14.transforms.scale.set(inst.size_08, inst.size_08, inst.size_08);
-
-        applyModelRotationAndScale(this.tmdDustModel_800d4d40);
-        this.renderSmapModel(this.tmdDustModel_800d4d40);
-
-        this.tmdDustModel_800d4d40.remainingFrames_9e = 0;
-        this.tmdDustModel_800d4d40.interpolationFrameIndex = 0;
-
-        this.tmdDustModel_800d4d40.modelParts_00[0].coord2_04.flg--;
-        inst.tick_00++;
-
-        prev = inst;
-        inst = inst.next_1c;
-      }
-      //LAB_800ef888
-    }
-    //LAB_800ef894
-  }
-
-  @Method(0x800ef8acL)
-  private void renderOrthoQuadTrailEffects() {
-    final int[] v = new int[4];
-    v[3] = 64; // Other values are 0
-
-    //LAB_800ef9cc
-    OrthoTrailParticle54 prev = this.orthoQuadTrail;
-    OrthoTrailParticle54 inst = prev.next_50;
-    while(inst != null) {
-      if(inst.tick_04 >= inst.maxTicks_06 * (3 - vsyncMode_8007a3b8)) {
-        prev.next_50 = inst.next_50;
-        inst = prev.next_50;
-      } else {
-        //LAB_800efa08
-        final GpuCommandPoly cmd = new GpuCommandPoly(4);
-
-        final int mode = inst.renderMode_00;
-        if(mode == 0 || mode == 2) {
-          //LAB_800efa44
-          final int offsetX = this.screenOffsetX_800cb568 - inst.x_18;
-          final int offsetY = this.screenOffsetY_800cb56c - inst.y_1c;
-
-          cmd
-            .pos(0, offsetX + inst.sxy0_20.x, offsetY + inst.sxy0_20.y)
-            .pos(1, offsetX + inst.sxy1_28.x, offsetY + inst.sxy1_28.y)
-            .pos(2, offsetX + inst.sxy2_30.x, offsetY + inst.sxy2_30.y)
-            .pos(3, offsetX + inst.sxy3_38.x, offsetY + inst.sxy3_38.y);
-
-          if(mode == 2) {
-            cmd
-              .clut(960, 464)
-              .vramPos(960, 256)
-              .bpp(Bpp.BITS_4)
-              .translucent(Translucency.B_MINUS_F);
-          } else {
-            //LAB_800efb64
-            cmd
-              .clut((this.cluts_800d6068[7] & 0b111111) * 16, this.cluts_800d6068[7] >>> 6)
-              .vramPos((this.texPages_800d6050[7] & 0b1111) * 64, (this.texPages_800d6050[7] & 0b10000) != 0 ? 256 : 0)
-              .translucent(Translucency.of(this.texPages_800d6050[7] >>> 5 & 0b11))
-              .bpp(Bpp.of(this.texPages_800d6050[7] >>> 7 & 0b11));
-          }
-        } else if(mode == 1) {
-          //LAB_800efb7c
-          inst.size_08 += inst.sizeStep_0c / (3 - vsyncMode_8007a3b8);
-          inst.sxy0_20.x = inst.z0_26 - inst.size_08 / 2.0f;
-          inst.sxy0_20.y = inst.z1_2e - inst.size_08 / 2.0f;
-          final float x = this.screenOffsetX_800cb568 - inst.x_18 + inst.sxy0_20.x;
-          final float y = this.screenOffsetY_800cb56c - inst.y_1c + inst.sxy0_20.y;
-
-          cmd
-            .pos(0, x, y)
-            .pos(1, x + inst.size_08, y)
-            .pos(2, x, y + inst.size_08)
-            .pos(3, x + inst.size_08, y + inst.size_08);
-
-          if((inst.tick_04 & 0x3) == 0) {
-            inst.z1_2e -= 1.0f / (3 - vsyncMode_8007a3b8);
-          }
-
-          //LAB_800efc4c
-          cmd
-            .clut((this.cluts_800d6068[6] & 0b111111) * 16, this.cluts_800d6068[6] >>> 6)
-            .vramPos((this.texPages_800d6050[6] & 0b1111) * 64, (this.texPages_800d6050[6] & 0b10000) != 0 ? 256 : 0)
-            .translucent(Translucency.of(this.texPages_800d6050[6] >>> 5 & 0b11))
-            .bpp(Bpp.of(this.texPages_800d6050[6] >>> 7 & 0b11));
-        }
-
-        //LAB_800efc64
-        if(inst.tick_04 >= this.particleFadeDelay_800d6c0c[inst.renderMode_00] * (3 - vsyncMode_8007a3b8)) {
-          inst.brightness_48 -= inst.stepBrightness_40 / (3 - vsyncMode_8007a3b8);
-
-          if(inst.brightness_48 >= 1.0f || inst.brightness_48 < 0.0f) {
-            inst.brightness_48 = 0.0f;
-          }
-        }
-
-        //LAB_800efcb8
-        cmd
-          .monochrome(inst.brightness_48)
-          .uv(0, this.orthoDustUs_800d6bdc[inst.textureIndex_02], v[inst.textureIndex_02])
-          .uv(1, this.orthoDustUs_800d6bdc[inst.textureIndex_02] + this.dustTextureWidths_800d6bec[inst.textureIndex_02], v[inst.textureIndex_02])
-          .uv(2, this.orthoDustUs_800d6bdc[inst.textureIndex_02], v[inst.textureIndex_02] + this.dustTextureHeights_800d6bfc[inst.textureIndex_02])
-          .uv(3, this.orthoDustUs_800d6bdc[inst.textureIndex_02] + this.dustTextureWidths_800d6bec[inst.textureIndex_02], v[inst.textureIndex_02] + this.dustTextureHeights_800d6bfc[inst.textureIndex_02]);
-
-        GPU.queueCommand(inst.z_4c, cmd);
-
-        inst.tick_04++;
-        prev = inst;
-        inst = inst.next_50;
-      }
-      //LAB_800efe48
-    }
-    //LAB_800efe54
-  }
-
-  @Method(0x800f0370L)
-  private void initAttachedSobjEffects() {
-    initModel(this.tmdDustModel_800d4d40, this.dustTmd, this.dustAnimation);
-    this.orthoQuadTrail.next_50 = null;
-    this.tmdTrail_800d4ec0.next_1c = null;
-    this.initLawPodTrail();
-  }
-
-  @Method(0x800f03c0L)
-  private TmdTrailParticle20 addTmdDustParticle(final TmdTrailParticle20 parent) {
-    final TmdTrailParticle20 child = new TmdTrailParticle20();
-    child.next_1c = parent.next_1c;
-    parent.next_1c = child;
-    return child;
-  }
-
-  @Method(0x800f0400L)
-  private OrthoTrailParticle54 addOrthoQuadTrailParticle(final OrthoTrailParticle54 parent) {
-    final OrthoTrailParticle54 child = new OrthoTrailParticle54();
-    child.next_50 = parent.next_50;
-    parent.next_50 = child;
-    return child;
-  }
-
-  @Method(0x800f0440L)
-  private void deallocateAttachedSobjEffects() {
-    this.deallocateTmdTrail();
-    this.deallocateOrthoQuadTrail();
-    this.deallocateLawPodTrail();
-  }
-
-  @Method(0x800f047cL)
-  private void renderAttachedSobjEffects() {
-    this.renderTmdTrail();
-    this.renderOrthoQuadTrailEffects();
-    this.renderLawPodTrail();
-  }
-
   @Method(0x800f04acL)
   private void initAttachedSobjEffectData(final AttachedSobjEffectData40 data) {
     data.tick_00 = 0;
@@ -6510,228 +6163,16 @@ public class SMap extends EngineState {
     data.transfer_1e.zero();
     data.size_28 = 0;
     data.oldFootprintInstantiationInterval_2c = 0;
-    data.instantiationIntervalDust30 = 0;
+    data.instantiationIntervalDust_30 = 0;
     data.instantiationIntervalFootprints_34 = 0;
     data.maxTicks_38 = 0;
     data.trailData_3c = null;
   }
 
-  @Method(0x800f058cL)
-  private void deallocateTmdTrail() {
-    final TmdTrailParticle20 prev = this.tmdTrail_800d4ec0;
-
-    if(prev.next_1c != null) {
-      //LAB_800f05b4
-      TmdTrailParticle20 next;
-      do {
-        final TmdTrailParticle20 inst = prev.next_1c;
-        next = inst.next_1c;
-        prev.next_1c = next;
-      } while(next != null);
-    }
-
-    //LAB_800f05d4
-  }
-
-  @Method(0x800f05e8L)
-  private void deallocateOrthoQuadTrail() {
-    final OrthoTrailParticle54 prev = this.orthoQuadTrail;
-
-    if(prev.next_50 != null) {
-      //LAB_800f0610
-      OrthoTrailParticle54 next;
-      do {
-        final OrthoTrailParticle54 inst = prev.next_50;
-        next = inst.next_50;
-        prev.next_50 = next;
-      } while(next != null);
-    }
-    //LAB_800f0630
-  }
-
-  @Method(0x800f0644L)
-  private void tickLawPodTrail(final Model124 model, final AttachedSobjEffectData40 data) {
-    if((data.tick_00 & (3 - vsyncMode_8007a3b8)) == 0) {
-      final LawPodTrailData18 trailData = data.trailData_3c;
-
-      if(trailData.countSegments_01 < trailData.maxCountSegments_00) {
-        final LawPodTrailSegment34 segment = new LawPodTrailSegment34();
-        TrailSegmentVertices14 newVerts = new TrailSegmentVertices14();
-
-        final MV transforms = new MV();
-        GsGetLs(model.coord2_14, transforms);
-
-        PushMatrix();
-        GTE.setTransforms(transforms);
-        GTE.perspectiveTransform(-trailData.width_08, 0.0f, 0.0f);
-        newVerts.vert0_00.x = GTE.getScreenX(2);
-        newVerts.vert0_00.y = GTE.getScreenY(2);
-        segment.z_20 = GTE.getScreenZ(3) / 4.0f;
-
-        GTE.perspectiveTransform(trailData.width_08, 0.0f, 0.0f);
-        newVerts.vert1_08.x = GTE.getScreenX(2);
-        newVerts.vert1_08.y = GTE.getScreenY(2);
-        segment.z_20 = GTE.getScreenZ(3) / 4.0f;
-        PopMatrix();
-
-        segment.tick_00 = 0;
-        segment.tpage_04 = GetTPage(Bpp.BITS_4, Translucency.of(trailData.translucency_0c), 972, 320);
-        this.initLawPodTrailSegmentColour(trailData, segment);
-        newVerts.vert0_00.x -= this.screenOffsetX_800cb568;
-        newVerts.vert0_00.y -= this.screenOffsetY_800cb56c;
-        newVerts.vert1_08.x -= this.screenOffsetX_800cb568;
-        newVerts.vert1_08.y -= this.screenOffsetY_800cb56c;
-
-        if(trailData.countSegments_01 == 0) {
-          trailData.currSegmentOriginVerts_14 = newVerts;
-          newVerts = new TrailSegmentVertices14();
-          newVerts.vert0_00.set(trailData.currSegmentOriginVerts_14.vert0_00);
-          newVerts.vert1_08.set(trailData.currSegmentOriginVerts_14.vert1_08);
-        }
-
-        //LAB_800f0928
-        segment.endpointVerts23_28 = newVerts;
-        segment.originVerts01_24 = trailData.currSegmentOriginVerts_14;
-        trailData.currSegmentOriginVerts_14 = newVerts;
-        segment.trailData_2c = trailData;
-        this.lawPodTrail_800d4f90.add(segment);
-        this.lawPodTrailVerts_800d4fd0.add(newVerts);
-        trailData.countSegments_01++;
-      }
-
-      data.transfer_1e.set(model.coord2_14.coord.transfer);
-    }
-    //LAB_800f094c
-  }
-
-  @Method(0x800f0970L)
-  private void renderLawPodTrail() {
-    //LAB_800f09c0
-    for(int i = 0; i < this.lawPodTrail_800d4f90.size(); i++) {
-      final LawPodTrailSegment34 segment = this.lawPodTrail_800d4f90.get(i);
-      final LawPodTrailData18 trailData = segment.trailData_2c;
-      if(segment.tick_00 <= trailData.maxTicks_06) {
-        final int tpage = segment.tpage_04;
-
-        //LAB_800f0b04
-        final GpuCommandPoly cmd = new GpuCommandPoly(4)
-          .translucent(Translucency.of(tpage >>> 5 & 0b11))
-          .pos(0, this.screenOffsetX_800cb568 + segment.originVerts01_24.vert0_00.x, this.screenOffsetY_800cb56c + segment.originVerts01_24.vert0_00.y)
-          .pos(1, this.screenOffsetX_800cb568 + segment.originVerts01_24.vert1_08.x, this.screenOffsetY_800cb56c + segment.originVerts01_24.vert1_08.y)
-          .pos(2, this.screenOffsetX_800cb568 + segment.endpointVerts23_28.vert0_00.x, this.screenOffsetY_800cb56c + segment.endpointVerts23_28.vert0_00.y)
-          .pos(3, this.screenOffsetX_800cb568 + segment.endpointVerts23_28.vert1_08.x, this.screenOffsetY_800cb56c + segment.endpointVerts23_28.vert1_08.y);
-
-        final Vector3f colour = new Vector3f();
-        if(segment.tick_00 > trailData.fadeDelay_02 - 1) {
-          //LAB_800f0d0c
-          segment.colour_14.sub(segment.colourStep_08);
-          segment.colour_14.x = Math.max(0, segment.colour_14.x);
-          segment.colour_14.y = Math.max(0, segment.colour_14.y);
-          segment.colour_14.z = Math.max(0, segment.colour_14.z);
-        }
-
-        //LAB_800f0d18
-        cmd.rgb((int)(segment.colour_14.x * 255), (int)(segment.colour_14.y * 255), (int)(segment.colour_14.z * 255));
-        GPU.queueCommand(segment.z_20, cmd);
-
-        segment.tick_00++;
-      } else {
-        trailData.countSegments_01--;
-        TrailSegmentVertices14 currSegmentOriginVerts = segment.originVerts01_24;
-
-        //LAB_800f09fc
-        for(int j = 0; j < this.lawPodTrailVerts_800d4fd0.size(); j++) {
-          if(this.lawPodTrailVerts_800d4fd0.get(j) == currSegmentOriginVerts) {
-            //LAB_800f0ae8
-            this.lawPodTrailVerts_800d4fd0.remove(j);
-            break;
-          }
-        }
-
-        //LAB_800f0a18
-        if(trailData.countSegments_01 == 0) {
-          currSegmentOriginVerts = segment.endpointVerts23_28;
-
-          //LAB_800f0a38
-          for(int j = 0; j < this.lawPodTrailVerts_800d4fd0.size(); j++) {
-            if(this.lawPodTrailVerts_800d4fd0.get(j) == currSegmentOriginVerts) {
-              //LAB_800f0acc
-              this.lawPodTrailVerts_800d4fd0.remove(j);
-              break;
-            }
-          }
-          //LAB_800f0a54
-        }
-
-        //LAB_800f0a58
-        if(trailData.countSegments_01 == 0) {
-          if(this.lawPodTrail_800d4f90.size() == 2) {
-            //LAB_800f0aa4
-            this.lawPodTrailVerts_800d4fd0.clear();
-          }
-        }
-      }
-      //LAB_800f0d68
-    }
-
-    //LAB_800f0d74
-    if(this.lawPodTrailSegmentCount_800f9e78 == 0 && this.lawPodTrail_800d4f90.size() == 1) {
-      //LAB_800f0da8
-      this.lawPodTrailVerts_800d4fd0.clear();
-    }
-    //LAB_800f0dc8
-  }
-
-  @Method(0x800f0df0L)
-  private void initLawPodTrailSegmentColour(final LawPodTrailData18 trailData, final LawPodTrailSegment34 segment) {
-    segment.colour_14.set(trailData.r_10 / 255.0f, trailData.g_11 / 255.0f, trailData.b_12 / 255.0f);
-    segment.colourStep_08.set(segment.colour_14).div(trailData.countFadeSteps_04);
-  }
-
-  @Method(0x800f0e60L)
-  private void initLawPodTrail() {
-    this.lawPodTrailSegmentCount_800f9e78 = 0;
-  }
-
-  @Method(0x800f0e7cL)
-  private void deallocateLawPodTrail() {
-    this.lawPodTrail_800d4f90.clear();
-
-    //LAB_800f0ec8
-    //LAB_800f0edc
-    this.lawPodTrailVerts_800d4fd0.clear();
-
-    //LAB_800f0efc
-    this.clearLawPodTrailSegmentsList();
-    this.lawPodTrailSegmentCount_800f9e78 = 0;
-  }
-
-  @Method(0x800f0f20L)
-  private LawPodTrailData18 addLawPodTrailSegment() {
-    LawPodTrailData18 segment = null;
-
-    //LAB_800f0f3c
-    for(int i = 0; i < 8; i++) {
-      if(this.lawPodTrailSegments_800f9e7c[i] == null) {
-        segment = new LawPodTrailData18();
-        this.lawPodTrailSegments_800f9e7c[i] = segment;
-        break;
-      }
-    }
-
-    //LAB_800f0f78
-    return segment;
-  }
-
-  @Method(0x800f0fe8L)
-  private void clearLawPodTrailSegmentsList() {
-    for(int i = 0; i < 8; i++) {
-      if(this.lawPodTrailSegments_800f9e7c[i] != null) {
-        this.lawPodTrailSegments_800f9e7c[i] = null;
-        this.lawPodTrailSegmentCount_800f9e78--;
-      }
-    }
-  }
+  //@Method(0x800f0e60L)
+  //private void initLawPodTrail() {
+  //  this.lawPodTrailSegmentCount_800f9e78 = 0;
+  //}
 
   @ScriptDescription("Allocates/initializes static struct containing smoke plume particle data.")
   @ScriptParam(direction = ScriptParam.Direction.IN, type = ScriptParam.Type.INT_ARRAY, name = "smokeData", description = "An array of data for the smoke plume particle data struct")
@@ -6791,32 +6232,7 @@ public class SMap extends EngineState {
   @ScriptParam(direction = ScriptParam.Direction.IN, type = ScriptParam.Type.INT, name = "b")
   @Method(0x800f1634L)
   private FlowControl scriptInitLawPodTrail(final RunningScript<?> script) {
-    final ScriptState<?> state = script.scriptState_04;
-
-    script.params_20[9] = new ScriptStorageParam(state, 0);
-
-    final SubmapObject210 sobj = (SubmapObject210)scriptStatePtrArr_800bc1c0[state.storage_44[0]].innerStruct_00;
-    if(script.params_20[0].get() == 0 || this.lawPodTrailSegmentCount_800f9e78 >= 8) {
-      //LAB_800f1698
-      sobj.attachedEffectData_1d0.shouldRenderLawPodTrail_18 = false;
-    } else {
-      //LAB_800f16a4
-      final LawPodTrailData18 segment = this.addLawPodTrailSegment();
-      segment.maxCountSegments_00 = script.params_20[1].get();
-      segment.countSegments_01 = 0;
-      segment.fadeDelay_02 = script.params_20[2].get();
-      segment.countFadeSteps_04 = script.params_20[3].get();
-      segment.maxTicks_06 = segment.fadeDelay_02 + segment.countFadeSteps_04;
-      segment.width_08 = script.params_20[4].get();
-      segment.translucency_0c = script.params_20[5].get();
-      segment.r_10 = script.params_20[6].get();
-      segment.g_11 = script.params_20[7].get();
-      segment.b_12 = script.params_20[8].get();
-      segment.currSegmentOriginVerts_14 = null;
-      sobj.attachedEffectData_1d0.shouldRenderLawPodTrail_18 = script.params_20[0].get() == 1;
-      sobj.attachedEffectData_1d0.trailData_3c = segment;
-      this.lawPodTrailSegmentCount_800f9e78++;
-    }
+    this.attachedSobjEffect.initLawPodTrail(script);
 
     //LAB_800f1784
     return FlowControl.CONTINUE;
@@ -6983,7 +6399,7 @@ public class SMap extends EngineState {
     final SubmapObject210 sobj = SCRIPTS.getObject(script.params_20[0].get(), SubmapObject210.class);
     sobj.attachedEffectData_1d0.shouldRenderTmdDust_04 = script.params_20[1].get() == 1;
     sobj.attachedEffectData_1d0.size_28 = script.params_20[2].get();
-    sobj.attachedEffectData_1d0.instantiationIntervalDust30 = script.params_20[3].get();
+    sobj.attachedEffectData_1d0.instantiationIntervalDust_30 = script.params_20[3].get();
 
     if(script.params_20[4].get() == 0) {
       sobj.attachedEffectData_1d0.maxTicks_38 = 1;
@@ -6995,7 +6411,7 @@ public class SMap extends EngineState {
     if(!sobj.attachedEffectData_1d0.shouldRenderTmdDust_04) {
       sobj.attachedEffectData_1d0.transfer_1e.zero();
       sobj.attachedEffectData_1d0.size_28 = 1;
-      sobj.attachedEffectData_1d0.instantiationIntervalDust30 = 0;
+      sobj.attachedEffectData_1d0.instantiationIntervalDust_30 = 0;
       sobj.attachedEffectData_1d0.maxTicks_38 = 0;
     }
 
@@ -7069,6 +6485,8 @@ public class SMap extends EngineState {
       sobj2.attachedEffectData_1d0.shouldRenderFootprints_08 = true;
       sobj2.attachedEffectData_1d0.footprintMode_10 = 0;
       sobj2.attachedEffectData_1d0.instantiationIntervalFootprints_34 = 9;
+
+      this.attachedSobjEffect.initOrthoQuadTrail(150, sobj2.attachedEffectData_1d0.instantiationIntervalFootprints_34);
     } else {
       //LAB_800f22b8
       sobj2.attachedEffectData_1d0.shouldRenderFootprints_08 = false;
@@ -7078,7 +6496,7 @@ public class SMap extends EngineState {
     return FlowControl.CONTINUE;
   }
 
-  @ScriptDescription("Initializes footprint attached sobj effect. 0 = individual, 1 = skid")
+  @ScriptDescription("Re-initializes footprint attached sobj effect. 0 = individual, 1 = skid")
   @ScriptParam(direction = ScriptParam.Direction.IN, type = ScriptParam.Type.INT, name = "footprintMode", description = "Style of footprint (individual or skid).")
   @ScriptParam(direction = ScriptParam.Direction.IN, type = ScriptParam.Type.INT, name = "instantiationTicks", description = "New number of ticks before a new particle is instantiated.")
   @ScriptParam(direction = ScriptParam.Direction.IN, type = ScriptParam.Type.INT, name = "sobjIndex", description = "The SubmapObject210 script index")
@@ -7099,6 +6517,8 @@ public class SMap extends EngineState {
       } else if(newInterval == 1) {
         sobj.attachedEffectData_1d0.shouldRenderFootprints_08 = true;
         sobj.attachedEffectData_1d0.instantiationIntervalFootprints_34 = sobj.attachedEffectData_1d0.oldFootprintInstantiationInterval_2c;
+
+        this.attachedSobjEffect.initOrthoQuadTrail(150, sobj.attachedEffectData_1d0.instantiationIntervalFootprints_34);
       }
       //LAB_800f2340
     } else if(footprintMode == 1) {
@@ -7109,6 +6529,8 @@ public class SMap extends EngineState {
       if(sobj.attachedEffectData_1d0.instantiationIntervalFootprints_34 == 0) {
         sobj.attachedEffectData_1d0.instantiationIntervalFootprints_34 = 1;
       }
+
+      this.attachedSobjEffect.initOrthoQuadTrail(150, sobj.attachedEffectData_1d0.instantiationIntervalFootprints_34);
     }
 
     //LAB_800f2374
@@ -7130,6 +6552,8 @@ public class SMap extends EngineState {
     final SubmapObject210 sobj = (SubmapObject210)scriptStatePtrArr_800bc1c0[script.scriptState_04.storage_44[0]].innerStruct_00;
     sobj.attachedEffectData_1d0.instantiationIntervalFootprints_34 = Math.max(1, script.params_20[0].get());
 
+    this.attachedSobjEffect.initOrthoQuadTrail(150, sobj.attachedEffectData_1d0.instantiationIntervalFootprints_34);
+
     //LAB_800f23e4
     return FlowControl.CONTINUE;
   }
@@ -7149,7 +6573,7 @@ public class SMap extends EngineState {
       //LAB_800f2430
       sobj.attachedEffectData_1d0.shouldRenderOrthoDust_0c = true;
       sobj.attachedEffectData_1d0.size_28 = script.params_20[1].get();
-      sobj.attachedEffectData_1d0.instantiationIntervalDust30 = script.params_20[2].get();
+      sobj.attachedEffectData_1d0.instantiationIntervalDust_30 = script.params_20[2].get();
       sobj.attachedEffectData_1d0.maxTicks_38 = Math.max(1, script.params_20[3].get());
     } else {
       //LAB_800f2484
@@ -7206,14 +6630,7 @@ public class SMap extends EngineState {
   @ScriptParam(direction = ScriptParam.Direction.IN, type = ScriptParam.Type.BOOL, name = "shouldDeallocate", description = "Deallocate trail effect when true.")
   @Method(0x800f25a8L)
   private FlowControl scriptDeallocateLawPodTrail(final RunningScript<?> script) {
-    final ScriptState<?> state = script.scriptState_04;
-    script.params_20[1] = new ScriptStorageParam(state, 0);
-
-    if(script.params_20[0].get() == 1) {
-      this.deallocateLawPodTrail();
-      final SubmapObject210 sobj = (SubmapObject210)scriptStatePtrArr_800bc1c0[state.storage_44[0]].innerStruct_00;
-      sobj.attachedEffectData_1d0.shouldRenderLawPodTrail_18 = false;
-    }
+    this.attachedSobjEffect.deallocateLawPodTrail(script);
 
     //LAB_800f2604
     return FlowControl.CONTINUE;
@@ -7767,7 +7184,7 @@ public class SMap extends EngineState {
       //LAB_800f4714
       this.deallocateSavePoint();
       this.deallocateSmokeAndSnow();
-      this.deallocateAttachedSobjEffects();
+      this.attachedSobjEffect.deallocateAttachedSobjEffects();
       this.submapEffectsLoadMode_800f9ea8 = 0;
       this.submapEffectsState_800f9eac = 0;
       return;
@@ -7787,7 +7204,7 @@ public class SMap extends EngineState {
       this.smokePlumeEffect_800d5fd8.effectShouldRender = false;
       this.initTriangleIndicators();
       this.initSavePoint();
-      this.initAttachedSobjEffects();
+      this.attachedSobjEffect.initAttachedSobjEffects();
       this.submapEffectsLoadMode_800f9ea8++;
       this.submapEffectsState_800f9eac = 2;
     }
